+++ src/plugins/nsfe_extractor.c
@@ -175,7 +175,7 @@ info_extract (struct EXTRACTOR_ExtractContext *ec,
 
   if (size < 8)
     return 0;
-  if (size >
+  if ((ssize_t) size >
       ec->read (ec->cls,
 		&data,
 		size))
@@ -243,7 +243,7 @@ tlbl_extract (struct EXTRACTOR_ExtractContext *ec,
   void *data;
   const char *cdata;
 
-  if (size >
+  if ((ssize_t) size >
       ec->read (ec->cls,
 		&data,
 		size))
@@ -285,7 +285,7 @@ auth_extract (struct EXTRACTOR_ExtractContext *ec,
 
   if (left < 1)
     return 0;
-  if (size >
+  if ((ssize_t) size >
       ec->read (ec->cls,
 		&data,
 		size))
@@ -342,7 +342,7 @@ EXTRACTOR_nsfe_extract_method (struct EXTRACTOR_ExtractContext *ec)
   uint32_t chunksize;
   int ret;
 
-  if (sizeof (struct header) >
+  if ((ssize_t) sizeof (struct header) >
       ec->read (ec->cls,
 		&data,
 		sizeof (struct header)))
